+++ packages/ui/Scrollable/ScrollableNative.js
@@ -1259,12 +1259,12 @@ class ScrollableBaseNative extends Component {
 
 		if (containerRef.current && containerRef.current.addEventListener) {
 			containerRef.current.addEventListener('wheel', this.onWheel);
+			containerRef.current.addEventListener('mousedown', this.onMouseDown);
 		}
 
 		if (childRefCurrent.containerRef.current) {
 			if (childRefCurrent.containerRef.current.addEventListener) {
 				childRefCurrent.containerRef.current.addEventListener('scroll', this.onScroll, {capture: true, passive: true});
-				childRefCurrent.containerRef.current.addEventListener('mousedown', this.onMouseDown);
 			}
 			this.childRefCurrent.containerRef.current.style.scrollBehavior = 'smooth';
 		}
@@ -1284,11 +1284,11 @@ class ScrollableBaseNative extends Component {
 
 		if (containerRef.current && containerRef.current.removeEventListener) {
 			containerRef.current.removeEventListener('wheel', this.onWheel);
+			containerRef.current.removeEventListener('mousedown', this.onMouseDown);
 		}
 
 		if (childRefCurrent.containerRef.current && childRefCurrent.containerRef.current.removeEventListener) {
 			childRefCurrent.containerRef.current.removeEventListener('scroll', this.onScroll, {capture: true, passive: true});
-			childRefCurrent.containerRef.current.removeEventListener('mousedown', this.onMouseDown);
 		}
 
 		if (this.props.removeEventListeners) {