+++ packages/ui/CHANGELOG.md
@@ -4,12 +4,13 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ## [unreleased]
 
-### Fixed
-
-- `ui/Scroller`, `ui/VirtualList.VirtualGridList`, and `ui/VirtualList.VirtualList` to suppress spotlight on scrollbars when `spotlightDisabled`
 ### Added
 
 - `ui/Icon`, `ui/IconButton`, and `ui/LabeledIcon` prop `flip` to flip the icon horizontally, vertically, or both
+
+### Fixed
+
+- `ui/Scroller`, `ui/VirtualList.VirtualGridList`, and `ui/VirtualList.VirtualList` to suppress spotlight on scrollbars when `spotlightDisabled`
 
 ## [3.0.0-beta.2] - 2019-07-23